refactor: use consistent naming
