+++ pages/personalization/hooks/use-user-preferences.js
@@ -1,5 +1,5 @@
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { useVisitorData } from '../../../shared/client/useVisitorData';
+import { useVisitorData } from '../../../shared/client/use-visitor-data';
 
 const GET_USER_PREFERENCES_QUERY = 'GET_USER_PREFERENCES_QUERY';
